создание компонента UserCart
