+++ src/components/UserCart/index.js
@@ -1,0 +1,2 @@
+
+export { UserCart } from './UserCart';
