+++ src/components/UserCart/style.css
@@ -1,0 +1,6 @@
+
+/* USERCART */
+
+.cart__list {
+    margin: 30px;
+}
