+++ src/pages/Cart/Cart.js
@@ -1,14 +1,14 @@
 
 import './style.css';
 import React from 'react';
-import { removeFromCart, clearCart } from '../../store/cartSlice';
 import { useDispatch, useSelector } from 'react-redux';
 import { NavLink } from 'react-router-dom';
+import { clearCart } from '../../store/cartSlice';
 import { Button } from '@mui/material';
+import { UserCart } from './../../components/UserCart';
 
 function Cart () {
     const dispatch = useDispatch();
-    const cartItems = useSelector(state => state.cart.cart);
     const cartTotalPrice = useSelector(state => state.cart.totalPrice);
     const logIn = localStorage.getItem('logIn');
     const userName = logIn === null ? 'please, log in' : logIn; 
@@ -18,20 +18,6 @@ function Cart () {
         dispatch(clearCart());
     }
 
-    const userCart = cartItems.map(([id, title, price]) => {
-        return <li key={id}>
-                    <div className="cart__item">
-                        <p>
-                            {title + ': ' + price + '$'}
-                        </p>
-                        <Button onClick={() => dispatch(removeFromCart([id, price]))}
-                         sx={{color: '#802580'}}>
-                            DELETE FROM CART
-                        </Button>
-                    </div>
-                </li>
-    });
-
     return (
         <div className="cart wrapper">
             <div className="cart__info">
@@ -40,9 +26,7 @@ function Cart () {
                     <p>{userName}</p>
                 </div>
 
-                <ol className="cart__list">
-                    {userCart}
-                </ol>
+                <UserCart />
 
                 <div className="cart__order">
                     <div className="cart__price">
